refactor: movement 패키지 아래로 direction 및 policy 이동
