+++ src/main/java/chess/domain/Movement.java
@@ -1,0 +1,4 @@
+package chess.domain;
+
+public class Movement {
+}
