--- src/main/java/chess/domain/direction/Direction.java
@@ -1,9 +1,0 @@
-package chess.domain.direction;
-
-import chess.domain.Position;
-import java.util.List;
-
-public interface Direction {
-
-    boolean canReach(Position from, Position to, List<Position> obstacle);
-}
