--- src/main/java/chess/domain/direction/DownRightDirection.java
@@ -1,19 +1,0 @@
-package chess.domain.direction;
-
-import chess.domain.Position;
-
-public class DownRightDirection extends StraightDirection {
-
-    DownRightDirection(final int moveCount) {
-        super(moveCount);
-    }
-
-    @Override
-    Position next(final Position from) {
-        if (from.isMinimumRow() || from.isMaximumColumn()) {
-            return from;
-        }
-
-        return new Position(from.row() - 1, from.column() + 1);
-    }
-}
