--- src/main/java/chess/domain/direction/UpLeftDirection.java
@@ -1,18 +1,0 @@
-package chess.domain.direction;
-
-import chess.domain.Position;
-
-public class UpLeftDirection extends StraightDirection {
-
-    UpLeftDirection(final int moveCount) {
-        super(moveCount);
-    }
-
-    @Override
-    Position next(final Position from) {
-        if (from.isMaximumRow() || from.isMinimumColumn()) {
-            return from;
-        }
-        return new Position(from.row() + 1, from.column() - 1);
-    }
-}
