--- src/main/java/chess/domain/direction/UpRightDirection.java
@@ -1,18 +1,0 @@
-package chess.domain.direction;
-
-import chess.domain.Position;
-
-public class UpRightDirection extends StraightDirection {
-
-    public UpRightDirection(final int moveCount) {
-        super(moveCount);
-    }
-
-    @Override
-    Position next(final Position from) {
-        if (from.isMaximumColumn() || from.isMaximumRow()) {
-            return from;
-        }
-        return new Position(from.row() + 1, from.column() + 1);
-    }
-}
