+++ src/main/java/chess/domain/movement/direction/Direction.java
@@ -1,0 +1,9 @@
+package chess.domain.movement.direction;
+
+import chess.domain.Position;
+import java.util.List;
+
+public interface Direction {
+
+    boolean canReach(Position from, Position to, List<Position> obstacle);
+}
