+++ src/main/java/chess/domain/movement/direction/DownDirection.java
@@ -1,0 +1,18 @@
+package chess.domain.movement.direction;
+
+import chess.domain.Position;
+
+public class DownDirection extends StraightDirection {
+
+    DownDirection(final int moveCount) {
+        super(moveCount);
+    }
+
+    @Override
+    Position next(final Position from) {
+        if (from.isMinimumRow()) {
+            return from;
+        }
+        return new Position(from.row() - 1, from.column());
+    }
+}
