+++ src/main/java/chess/domain/movement/policy/CombinationPolicy.java
@@ -1,0 +1,19 @@
+package chess.domain.movement.policy;
+
+import chess.domain.Color;
+import java.util.List;
+
+public class CombinationPolicy implements Policy {
+
+    List<Policy> policies;
+
+    public CombinationPolicy(final Policy... policies) {
+        this.policies = List.of(policies);
+    }
+
+    @Override
+    public boolean isSatisfied(final Color color, final boolean firstMove, final boolean existEnemy) {
+        return policies.stream()
+                .allMatch(policy -> policy.isSatisfied(color, firstMove, existEnemy));
+    }
+}
