+++ src/main/java/chess/domain/movement/policy/EnemyExistPolicy.java
@@ -1,0 +1,11 @@
+package chess.domain.movement.policy;
+
+import chess.domain.Color;
+
+public class EnemyExistPolicy implements Policy {
+
+    @Override
+    public boolean isSatisfied(final Color color, final boolean firstMove, final boolean existEnemy) {
+        return existEnemy;
+    }
+}
