+++ src/main/java/chess/domain/movement/policy/FirstMovePolicy.java
@@ -1,0 +1,12 @@
+package chess.domain.movement.policy;
+
+import chess.domain.Color;
+
+
+public class FirstMovePolicy implements Policy {
+
+    @Override
+    public boolean isSatisfied(final Color color, final boolean firstMove, final boolean existEnemy) {
+        return firstMove;
+    }
+}
