+++ src/main/java/chess/domain/movement/policy/Policy.java
@@ -1,0 +1,7 @@
+package chess.domain.movement.policy;
+
+import chess.domain.Color;
+
+public interface Policy {
+    boolean isSatisfied(Color color, boolean firstMove, boolean existEnemy);
+}
