--- src/main/java/chess/domain/policy/ColorPolicy.java
@@ -1,17 +1,0 @@
-package chess.domain.policy;
-
-import chess.domain.Color;
-
-public class ColorPolicy implements Policy {
-
-    private final Color color;
-
-    public ColorPolicy(final Color color) {
-        this.color = color;
-    }
-
-    @Override
-    public boolean isSatisfied(final Color color, final boolean firstMove, final boolean existEnemy) {
-        return this.color == color;
-    }
-}
