--- src/main/java/chess/domain/policy/CombinationPolicy.java
@@ -1,19 +1,0 @@
-package chess.domain.policy;
-
-import chess.domain.Color;
-import java.util.List;
-
-public class CombinationPolicy implements Policy {
-
-    List<Policy> policies;
-
-    public CombinationPolicy(final Policy... policies) {
-        this.policies = List.of(policies);
-    }
-
-    @Override
-    public boolean isSatisfied(final Color color, final boolean firstMove, final boolean existEnemy) {
-        return policies.stream()
-                .allMatch(policy -> policy.isSatisfied(color, firstMove, existEnemy));
-    }
-}
