--- src/main/java/chess/domain/policy/Policy.java
@@ -1,7 +1,0 @@
-package chess.domain.policy;
-
-import chess.domain.Color;
-
-public interface Policy {
-    boolean isSatisfied(Color color, boolean firstMove, boolean existEnemy);
-}
