+++ src/test/java/chess/domain/movement/policy/EnemyExistPolicyTest.java
@@ -1,0 +1,35 @@
+package chess.domain.movement.policy;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+import chess.domain.Color;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+
+class EnemyExistPolicyTest {
+
+    @ParameterizedTest
+    @EnumSource(Color.class)
+    @DisplayName("해당 위치의 적 유무 정책과 일치 할 경우 해당 정책을 만족한다.")
+    void isSatisfied(Color color) {
+        EnemyExistPolicy policy = new EnemyExistPolicy();
+        assertAll(
+                () -> assertThat(policy.isSatisfied(color, true, true)).isTrue(),
+                () -> assertThat(policy.isSatisfied(color, false, true)).isTrue()
+        );
+    }
+
+    @ParameterizedTest
+    @EnumSource(Color.class)
+    @DisplayName("해당 위치의 적 유무 정책과 일치하지 않을 경우 해당 정책을 만족하지 않는다.")
+    void isNotSatisfied(Color color) {
+        EnemyExistPolicy policy = new EnemyExistPolicy();
+
+        assertAll(
+                () -> assertThat(policy.isSatisfied(color, true, false)).isFalse(),
+                () -> assertThat(policy.isSatisfied(color, false, false)).isFalse()
+        );
+    }
+}
